remove linter errors
