+++ index.html
@@ -22,12 +22,12 @@
 
       <form class="bookform hidden mb-3" action="" id="bookform">
         <div class="form-group">
-          <label for="bookform">Book Title: </label>
+          <label for="booktitle">Book Title: </label>
           <input
             type="text"
             class="form-control"
             name="title"
-            id="title"
+            id="booktitle"
             placeholder="Enter book title"
           />
         </div>
@@ -57,7 +57,6 @@
         <tbody id="bookslist"></tbody>
       </table>
     </div>
-    <script src="js/books.js"></script>
-    <script src="js/index.js"></script>
+    <script type="module" src="/js/index.js"></script>
   </body>
 </html>
